fix Issue deserialization when flattening occurs
when using the derived deserialize impl quirks in the
serde_json::RawValue (or RawJson in this crate) will cause automatic
deserialization failure with a bizarre error when serde attempts to
invoke the flatten functionality.

To work around this, we manually implement the deserialization impl,
allowing us to side step the issue.